fix: Restore internals of strhash to compile on 32 bit architectures
Fixes 4212

+++ src/libutil/strutil_test.cpp
@@ -346,13 +346,13 @@ void
 test_hash()
 {
     using namespace Strutil;
-    OIIO_CHECK_EQUAL(strhash("foo"), 6150913649986995171);
-    OIIO_CHECK_EQUAL(strhash(std::string("foo")), 6150913649986995171);
-    OIIO_CHECK_EQUAL(strhash(string_view("foo")), 6150913649986995171);
+    OIIO_CHECK_EQUAL(strhash("foo"), size_t(6150913649986995171));
+    OIIO_CHECK_EQUAL(strhash(std::string("foo")), size_t(6150913649986995171));
+    OIIO_CHECK_EQUAL(strhash(string_view("foo")), size_t(6150913649986995171));
     OIIO_CHECK_EQUAL(strhash(""), 0);  // empty string hashes to 0
     // Check longer hash and ensure that it's really constexpr
     constexpr size_t hash = Strutil::strhash("much longer string");
-    OIIO_CHECK_EQUAL(hash, 16257490369375554819ULL);
+    OIIO_CHECK_EQUAL(hash, size_t(16257490369375554819ULL));
 }
